pyproject.toml
[project]
name = "datapulse"
version = "0.1.0"
description = "Personal AI career intelligence platform for data analysts and engineers"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv",
    "supabase",
    "feedparser",
    "anthropic",
    "streamlit>=1.45.0",
    "plotly>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]

[project.scripts]
datapulse = "datapulse.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
